+++ packages/web3-providers/src/entry-types.ts
@@ -11,6 +11,7 @@ export * from './types/Fuse.js'
 export * from './types/GasOption.js'
 export * from './types/History.js'
 export * from './types/Instagram.js'
+export * from './types/Owner.js'
 export * from './types/Mask.js'
 export * from './types/MaskX.js'
 export * from './types/Minds.js'
